feat(arena): guard orphan cleanup by process identity
Record boot and process-start identities before exposing new regions, then remove registry entries only after death or PID reuse is proven. Preserve ambiguous and hostile records, emit audit decisions, and provide native Windows and Linux probes with a compile-checked macOS backend.

+++ engine/src/arena/platform_unix.rs
@@ -19,7 +19,7 @@ impl Mapping {
             libc::shm_open(
                 os_name.as_ptr(),
                 libc::O_CREAT | libc::O_EXCL | libc::O_RDWR,
-                libc::S_IRUSR | libc::S_IWUSR,
+                (libc::S_IRUSR | libc::S_IWUSR) as libc::c_uint,
             )
         };
         if descriptor < 0 {
